requirements-optimized.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Vector Database & Embeddings
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
numpy>=1.26.0

# Cloud LLM Provider
groq>=0.4.0

# Document Processing - Core
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
PyYAML>=6.0

# Office Documents (Optional - comment out if not needed)
openpyxl>=3.1.0
xlrd>=2.0.0
python-pptx>=0.6.21

# Frontend
streamlit>=1.28.0

# HTTP & Utilities
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# OCR Support (Optional - uncomment if needed for image processing)
# pytesseract>=0.3.10
# Pillow>=10.0.0
# lxml>=4.9.0